requirements.txt
jinja2==3.1.6       # pyyaml dependency
lark==1.2.2         # pytest-cov dependency
numpy==1.26.4
obd==0.7.3
opencv-python==4.10.0.84
psutil==6.1.0
python-daemon==3.1.0
pytest==8.3.2
pytest-cov==5.0.0
pyyaml==6.0.2
typeguard==4.4.1    # pyyaml dependency
